+++ CHANGELOG.md
@@ -14,9 +14,17 @@ without compromising stable operations.
 
 ## Unreleased
 
+## 0.61.0
+
 - `python-memory`: make job option available on YARN, add a default config
+- `load_stac`: optimize resolution with regard to requested bands (experimental) ([#1043](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1043))
+- `load_stac`: apply offset (experimental) ([#1051](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1051))
+- Deprecate non-standard "logging-threshold" job option in favor of standardized "log_level" job creation parameter ([#909](https://github.com/Open-EO/openeo-geopyspark-driver/issues/909))
+- Fail fast on UDF dependency installation failure ([#1048](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1048))
+- `load_stac`: avoid OOM on global low-res assets ([#1055](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1055))
 
 ## 0.60.1
+
 - `load_stac`: restore spatial dimensions of data cube loaded from STAC Collection that lacks `cube:dimensions`. ([#1036](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1036))
 - `sar_backscatter`: report soft-errors fraction in usage metrics
 - `apply_neighborhood`: for t='P1D', add date to dataarray.attrs with key 't'